VocaCare-Backend-FastAPI/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
python-jose[cryptography]
bcrypt==4.0.1
